% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{chapter_connect}
\alias{chapter_connect}
\title{chapter connect}
\usage{
chapter_connect(args = list(user = "jr", pass = "jr-pass", host =
  "localhost", port = 5432, dbname = "test"))
}
\arguments{
\item{args}{a list of arguments to pass to the database connection}
}
\value{
A connection object
}
\description{
A function for returning a database connection intended to
be used at the start of each chapter. Also checks whether the
database is available first.
}
