% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{check_compose}
\alias{check_compose}
\title{check compose}
\usage{
check_compose(dir = ".")
}
\arguments{
\item{dir}{directory path to check}
}
\description{
checks for the existence of a docker compose file
}
