% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_compose.R
\name{copy_compose}
\alias{copy_compose}
\alias{start_compose}
\alias{stop_compose}
\alias{setup_compose}
\title{Copy compose}
\usage{
copy_compose(target = ".")

start_compose(target = ".")

stop_compose(target = ".")

setup_compose()
}
\arguments{
\item{target}{target path for compose file}
}
\description{
Copies the docker compose file to the local directory
}
