% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sample_data.R
\name{load_sample_data}
\alias{load_sample_data}
\title{load sample data}
\usage{
load_sample_data(con)
}
\arguments{
\item{con}{A connection to a database}
}
\description{
A function to insert the movies and diamonds data into a SQL connection
}
