% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{start_database}
\alias{start_database}
\title{start database}
\usage{
start_database(dir = ".")
}
\arguments{
\item{dir}{location from which to call docker-compose}
}
\description{
runs system to call docker-compose up
}
