% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{check_database_exists}
\alias{check_database_exists}
\title{Check database}
\usage{
check_database_exists(args = NULL, verbose = FALSE)
}
\arguments{
\item{args}{a list of arguments to pass to the database connection. If NULL
use default args supplied by \code{get_default_args}}

\item{verbose}{logical if TRUE should give more info on errors}
}
\value{
logical TRUE if database accessible
}
\description{
Checks whether a post gres database is accessible to the session
}
