% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{copy_compose}
\alias{copy_compose}
\title{Copy compose}
\usage{
copy_compose(target = ".")
}
\arguments{
\item{target}{target path for compose file}
}
\description{
Copies the docker compose file to the local directory
}
