% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_default_args}
\alias{get_default_args}
\alias{fill_database}
\title{Function for filling an empty database}
\usage{
get_default_args(args = NULL)

fill_database(args = NULL)
}
\arguments{
\item{args}{a list of arguments to pass to the database connection. If NULL
use default args supplied by \code{get_default_args}}
}
\description{
Function should only be used for setting up a database for building
the notes and slides
}
